+++ src/dictionary/DefaultEditDate.ts
@@ -1,14 +1,16 @@
 import DefaultEdit, { DefaultEditInitOption } from "./DefaultEdit"
-import DictionaryValue from "../lib/DictionaryValue"
+import DictionaryValue, { functionType } from "../lib/DictionaryValue"
 
 type disabledDateConfig = { start?: unknown, end?: unknown }
 type disabledTimeConfig = { start?: unknown, end?: unknown }
 
 export interface DefaultEditDateOption {
   format: string
+  showFormat: string
   hideClear: boolean
   time?: {
     format: string
+    showFormat: string
     defaultValue: string
   }
   disabledDate?: (value: unknown) => boolean
@@ -17,9 +19,11 @@ export interface DefaultEditDateOption {
 
 export interface PartialDefaultEditDateOption {
   format?: string
+  showFormat?: string
   hideClear?: boolean
   time?: {
     format?: string
+    showFormat?: string
     defaultValue?: string
   }
   disabledDate?: disabledDateConfig | ((value: unknown) => boolean)
@@ -33,8 +37,11 @@ export interface DefaultEditDateInitOption extends DefaultEditInitOption {
 
 class DefaultEditDate extends DefaultEdit{
   static $name = 'DefaultEditDate'
+  static $edit: undefined | ((value: undefined | string, format: string) => undefined | unknown)
+  static $post: undefined | ((value: undefined | unknown, format: string) => undefined | string)
   static $defaultOption = {
     format: 'YYYY-MM-DD',
+    formatWithTime: 'YYYY-MM-DD HH:mm:ss',
     hideClear: false,
     time: {
       format: 'HH:mm:ss',
@@ -57,14 +64,19 @@ class DefaultEditDate extends DefaultEdit{
     super(initOption, parent, modName)
     this.type = initOption.type
     const option = initOption.option || {}
-    const $defaultOption = (this.constructor as typeof DefaultEditDate).$defaultOption
+    const $constructor = (this.constructor as typeof DefaultEditDate)
+    const $defaultOption = $constructor.$defaultOption
+    const format = option.format || option.time ? $defaultOption.formatWithTime : $defaultOption.format
     this.$option = {
-      format: option.format || $defaultOption.format,
+      format: format,
+      showFormat: option.showFormat || format,
       hideClear: option.hideClear === undefined ? $defaultOption.hideClear : option.hideClear
     }
     if (option.time) {
+      const timeFormat = option.time.format || $defaultOption.time.format
       this.$option.time = {
-        format: option.time.format || $defaultOption.time.format,
+        format: timeFormat,
+        showFormat: option.time.showFormat || timeFormat,
         defaultValue: option.time.defaultValue || $defaultOption.time.defaultValue
       }
     }
@@ -77,6 +89,24 @@ class DefaultEditDate extends DefaultEdit{
       if (typeof option.disabledTime === 'object') {
         this.$option.disabledTime = $defaultOption.disabledTime(option.disabledTime)
       }
+    }
+    if (this.edit === undefined) {
+      this.edit = function(this: DefaultEditDate, value: string) {
+        if ($constructor.$edit) {
+          return $constructor.$edit(value, this.$option.format)
+        } else {
+          return value
+        }
+      } as functionType<unknown>
+    }
+    if (this.post === undefined) {
+      this.post = function(this: DefaultEditDate, value: string) {
+        if ($constructor.$post) {
+          return $constructor.$post(value, this.$option.format)
+        } else {
+          return value
+        }
+      } as functionType<unknown>
     }
   }
 }