+++ src/dictionary/DefaultEditDateRange.ts
@@ -1,15 +1,18 @@
 import DefaultEdit, { DefaultEditInitOption } from "./DefaultEdit"
-import DictionaryValue from "../lib/DictionaryValue"
+import DefaultEditDate from "./DefaultEditDate"
+import DictionaryValue, { functionType } from "../lib/DictionaryValue"
 
 type disabledDateConfig = { start?: unknown, end?: unknown }
 type disabledTimeConfig = { start?: unknown, end?: unknown }
 
 export interface DefaultEditDateRangeOption {
   format: string
+  showFormat: string
   separator: string
   hideClear: boolean
   time?: {
     format: string
+    showFormat: string
     defaultValue: string[]
   }
   disabledDate?: (value: unknown) => boolean
@@ -18,10 +21,12 @@ export interface DefaultEditDateRangeOption {
 
 export interface PartialDefaultEditDateRangeOption {
   format?: string
+  showFormat?: string
   separator?: string
   hideClear?: boolean
   time?: {
     format?: string
+    showFormat?: string
     defaultValue?: string[]
   }
   disabledDate?: disabledDateConfig | ((value: unknown) => boolean)
@@ -35,12 +40,15 @@ export interface DefaultEditDateRangeInitOption extends DefaultEditInitOption {
 
 class DefaultEditDateRange extends DefaultEdit{
   static $name = 'DefaultEditDateRange'
+  static $edit: undefined | ((value: undefined | string, format: string) => undefined | unknown)
+  static $post: undefined | ((value: undefined | unknown, format: string) => undefined | string)
   static $defaultOption = {
-    format: 'YYYY-MM-DD',
+    format: undefined as undefined | string,
+    formatWithTime: undefined as undefined | string,
     separator: '-',
-    hideClear: false,
+    hideClear: undefined as undefined | boolean,
     time: {
-      format: 'HH:mm:ss',
+      format: undefined as undefined | string,
       defaultValue: ['00:00:00', '23:59:59']
     },
     disabledDate (option: disabledDateConfig) {
@@ -60,15 +68,20 @@ class DefaultEditDateRange extends DefaultEdit{
     super(initOption, parent, modName)
     this.type = initOption.type
     const option = initOption.option || {}
-    const $defaultOption = (this.constructor as typeof DefaultEditDateRange).$defaultOption
+    const $constructor = (this.constructor as typeof DefaultEditDateRange)
+    const $defaultOption = $constructor.$defaultOption
+    const format = option.format || option.time ? ($defaultOption.formatWithTime || DefaultEditDate.$defaultOption.formatWithTime) : ($defaultOption.format || DefaultEditDate.$defaultOption.format)
     this.$option = {
-      format: option.format || $defaultOption.format,
+      format: format,
+      showFormat: option.showFormat || format,
       separator: option.separator || $defaultOption.separator,
-      hideClear: option.hideClear === undefined ? $defaultOption.hideClear : option.hideClear
+      hideClear: option.hideClear === undefined ? ($defaultOption.hideClear === undefined ? DefaultEditDate.$defaultOption.hideClear : $defaultOption.hideClear) : option.hideClear
     }
     if (option.time) {
+      const timeFormat = option.time.format || $defaultOption.time.format || DefaultEditDate.$defaultOption.time.format
       this.$option.time = {
-        format: option.time.format || $defaultOption.time.format,
+        format: timeFormat,
+        showFormat: option.time.showFormat || timeFormat,
         defaultValue: option.time.defaultValue || $defaultOption.time.defaultValue
       }
     }
@@ -81,6 +94,28 @@ class DefaultEditDateRange extends DefaultEdit{
       if (typeof option.disabledTime === 'object') {
         this.$option.disabledTime = $defaultOption.disabledTime(option.disabledTime)
       }
+    }
+    if (this.edit === undefined) {
+      this.edit = function(this: DefaultEditDateRange, value: string) {
+        if ($constructor.$edit) {
+          $constructor.$edit(value, this.$option.format)
+        } else if (DefaultEditDate.$edit) {
+          DefaultEditDate.$edit(value, this.$option.format)
+        } else {
+          return value
+        }
+      } as functionType<unknown>
+    }
+    if (this.post === undefined) {
+      this.post = function(this: DefaultEditDateRange, value: string) {
+        if ($constructor.$post) {
+          $constructor.$post(value, this.$option.format)
+        } else if (DefaultEditDate.$post) {
+          DefaultEditDate.$post(value, this.$option.format)
+        } else {
+          return value
+        }
+      } as functionType<unknown>
     }
   }
 }